Refactor ApiEvent (follow up to removing MATS telemetry) (#3048)
* Remove unused enums.

* Minor fixes.

* Remove unused properties from ApiEvent.

* In ApiEvent, update string format for bool properties.

* Use StringBuilder to build telemetry string in HttpTelemetryManager.

* Use enum types instead of int in ApiEvent.

* Update LogMetricsFromAuthResult to use StringBuilder

* Add perf test for AcquireTokenForClient and OBO call without cache.

+++ src/client/Microsoft.Identity.Client/Internal/Requests/Silent/CacheSilentStrategy.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Identity.Client.ApiConfig.Parameters;
@@ -58,11 +57,11 @@ namespace Microsoft.Identity.Client.Internal.Requests.Silent
                 }
                 else
                 {
-                    if (AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo != (int)CacheRefreshReason.Expired)
+                    if (AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo != CacheRefreshReason.Expired)
                     {
                         cacheInfoTelemetry = CacheRefreshReason.NoCachedAccessToken;
                     }
-                } 
+                }
             }
             else
             {
@@ -70,9 +69,9 @@ namespace Microsoft.Identity.Client.Internal.Requests.Silent
                 logger.Info("Skipped looking for an Access Token because ForceRefresh or Claims were set. ");
             }
 
-            if (AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo == (int)CacheRefreshReason.NotApplicable)
+            if (AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo == CacheRefreshReason.NotApplicable)
             {
-                AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo = (int)cacheInfoTelemetry;
+                AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo = cacheInfoTelemetry;
             }
 
             // No AT or AT neesd to be refreshed 
@@ -89,7 +88,7 @@ namespace Microsoft.Identity.Client.Internal.Requests.Silent
                     // may fire a request to get a new token in the background
                     if (shouldRefresh)
                     {
-                        AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo = (int)CacheRefreshReason.ProactivelyRefreshed;
+                        AuthenticationRequestParameters.RequestContext.ApiEvent.CacheInfo = CacheRefreshReason.ProactivelyRefreshed;
 
                         SilentRequestHelper.ProcessFetchInBackground(
                         cachedAccessTokenItem,
